requirements.txt

opencv-python
flask==1.0.2

scipy
Pillow
matplotlib
keras==2.1.1
h5py
imgaug
IPython[all]
jupyter

absl-py==1.0.0
alabaster==0.7.12
amqp==2.6.1
aniso8601==9.0.1
argh==0.26.2
argon2-cffi==21.1.0
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
billiard==3.5.0.5
bleach==4.1.0
cachetools==4.2.4
celery==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==35.0.0
cycler==0.11.0
Cython==0.29.24
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dnspython==1.16.0
docutils==0.17.1


flake8==4.0.1
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Login==0.4.1
flask-mongoengine==0.9.5
flask-restplus==0.12.1
Flask-SocketIO==3.3.2
Flask-WTF==1.0.0
fonttools==4.28.1
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-images-download==2.5.0
greenlet==1.1.2
grpcio==1.41.1
h11==0.12.0
h5py==3.6.0
idna==3.3
imageio==2.10.5
imagesize==1.3.0
imantics==0.1.9
imgaug==0.4.0
importlib-metadata==4.8.2
importlib-resources==5.4.0
imutils==0.5.4
ipdb==0.13.9
ipykernel==6.5.0
ipyparallel==8.0.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
Keras==2.1.1
kiwisolver==1.3.2
kombu==4.3.0
lxml==4.6.4
Markdown==3.3.5
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.3.1
mkl-service==2.4.0
mongoengine==0.23.1
more-itertools==8.11.0
natsort==8.0.0
nbclient==0.5.8
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
ninja==1.10.2.2
nose==1.3.7
notebook==6.4.6
oauthlib==3.1.1
opencv-python==4.5.4.58
outcome==1.1.0
packaging==21.2
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.10.0
pymongo==3.12.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==3.9.3
pytest-ordering==0.6
python-dateutil==2.8.2
python-engineio==4.3.0
python-socketio==3.1.2
pytorch-ranger==0.1.1
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.0
QtPy==1.11.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.2
seaborn==0.11.2
selenium==4.0.0
Send2Trash==1.8.0
setuptools-scm==6.3.2
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
Sphinx==4.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2021.11.2
toml==0.10.2
tomli==1.2.2
torch==1.10.0
torch-optimizer==0.3.0
torch-summary==1.4.5
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
vine==1.3.0
watchdog==0.8.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.2
wsproto==1.0.0
WTForms==3.0.0
zipp==3.6.0